# app/openapi_formatted.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MetaMaster",
    "description": "Media Management Web Tool - Organize and manage your media library",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "List Movies",
        "description": "List movies with advanced filtering and sorting.\n\n- **genre**: Filter by genre (case-insensitive, optional)\n- **min_rating**: Minimum rating 0-10 (optional)\n- **max_rating**: Maximum rating 0-10 (optional)\n- **year**: Release year (optional)\n- **sort_by**: Sort field - title, rating, year, date_added (default: title)\n- **limit**: Number of items per page (1-100, default: 10)\n- **skip**: Offset from start (default: 0)",
        "operationId": "list_movies_api_v1_movies_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by genre (case-insensitive)",
              "title": "Genre"
            },
            "description": "Filter by genre (case-insensitive)"
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 10.0,
              "minimum": 0.0,
              "description": "Minimum rating (0-10)",
              "title": "Min Rating"
            },
            "description": "Minimum rating (0-10)"
          },
          {
            "name": "max_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 10.0,
              "minimum": 0.0,
              "description": "Maximum rating (0-10)",
              "title": "Max Rating"
            },
            "description": "Maximum rating (0-10)"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 1800,
              "description": "Release year",
              "title": "Year"
            },
            "description": "Release year"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort by: title, rating, year, date_added",
              "default": "title",
              "title": "Sort By"
            },
            "description": "Sort by: title, rating, year, date_added"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset from start",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset from start"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Alternative page number",
              "title": "Page"
            },
            "description": "Alternative page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Alternative page size",
              "title": "Page Size"
            },
            "description": "Alternative page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMovieResponseWithFilters"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Create Movie",
        "description": "Create a new movie",
        "operationId": "create_movie_api_v1_movies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/movies/search": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Search Movies",
        "description": "Search movies by title.\n\n- **q**: Search query (required)\n- **limit**: Number of items per page (1-100, default: 10)\n- **offset**: Offset from start (default: 0)",
        "operationId": "search_movies_api_v1_movies_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Search query",
              "title": "Q"
            },
            "description": "Search query"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset from start",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset from start"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Alternative page number",
              "title": "Page"
            },
            "description": "Alternative page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Alternative page size",
              "title": "Page Size"
            },
            "description": "Alternative page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMovieResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/movies/popular": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Popular Movies",
        "operationId": "popular_movies_api_v1_movies_popular_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset from start",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset from start"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Alternative page number",
              "title": "Page"
            },
            "description": "Alternative page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Alternative page size",
              "title": "Page Size"
            },
            "description": "Alternative page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMovieResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/movies/top-rated": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Top Rated Movies",
        "operationId": "top_rated_movies_api_v1_movies_top_rated_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset from start",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset from start"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Alternative page number",
              "title": "Page"
            },
            "description": "Alternative page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Alternative page size",
              "title": "Page Size"
            },
            "description": "Alternative page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMovieResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/movies/{movie_id}/scan": {
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Scan Movie",
        "description": "Run FFprobe on the movie's file and return updated metadata.",
        "operationId": "scan_movie_api_v1_movies__movie_id__scan_post",
        "parameters": [
          {
            "name": "movie_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Movie Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/movies/{movie_id}/enrich": {
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Trigger Movie Enrichment",
        "description": "Manually trigger external enrichment for a movie.",
        "operationId": "trigger_movie_enrichment_api_v1_movies__movie_id__enrich_post",
        "parameters": [
          {
            "name": "movie_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Movie Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaginatedMovieResponseWithFilters": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MovieResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of movies"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of movies matching filters"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Items per page"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Offset from start"
          },
          "filters": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FilterMetadata"
              }
            ],
            "description": "Applied filters and metadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "filters"
        ],
        "title": "PaginatedMovieResponseWithFilters",
        "description": "Paginated response for movies with filter metadata",
        "example": {
          "filters": {
            "applied_filters": {
              "genre": "Drama",
              "min_rating": 7.0
            },
            "sort_by": "rating",
            "total_results": 42
          },
          "items": [],
          "limit": 10,
          "offset": 0,
          "total": 42
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}